Updating post-install script to not necessarily require the mexa64 wrapper to be built
git-svn-id: svn+ssh://svn.cm.aol.com/advertising/adlearn/gen1/trunk/lib/cpp/DataProxy@297802 adebfd7e-f233-4696-9a66-a9d0f0efb86a

+++ hcmset/scripts/post-install.sh
@@ -32,7 +32,9 @@ cp -r ${TOPDIR}/sample/ ${TOPDIR}/service && rm -rf ${TOPDIR}/sample
 
 echo "--> Installing core libs to $LIBDIR"
 mkdir -p $LIBDIR
-mv ${TOPDIR}/DataProxy.mexa64 ${LIBDIR}
+if [ -e ${TOPDIR}/DataProxy.mexa64 ]; then
+	mv ${TOPDIR}/DataProxy.mexa64 ${LIBDIR}
+fi
 mv ${TOPDIR}/libDataProxy.so.* ${LIBDIR}
 ln -sfn libDataProxy.so.${VERSION} ${LIBDIR}/libDataProxy.so.${MIN_VER}
 ln -sfn libDataProxy.so.${MIN_VER} ${LIBDIR}/libDataProxy.so.${MAJ_VER}
